+++ src/shared/components/entry/EntryDownloadButton.tsx
@@ -1,0 +1,25 @@
+import { Button, DownloadIcon } from 'franklin-sites';
+import lazy from '../../utils/lazy';
+
+const EntryDownloadComponent = lazy(
+  /* istanbul ignore next */
+  () => import(/* webpackChunkName: "entry-download" */ './EntryDownload')
+);
+
+const EntryDownloadButton = ({
+  handleToggle,
+}: {
+  handleToggle: () => void;
+}) => (
+  <Button
+    variant="tertiary"
+    onPointerOver={EntryDownloadComponent.preload}
+    onFocus={EntryDownloadComponent.preload}
+    onClick={handleToggle}
+  >
+    <DownloadIcon />
+    Download
+  </Button>
+);
+
+export default EntryDownloadButton;
